backend/requirements.txt
fastapi==0.103.0
uvicorn==0.23.1
pydantic==1.10.2
pytest==7.2.0
pillow==9.3.0
python-multipart
tensorflow
python-dotenv
firebase-admin
cryptography
pycryptodome
